+++ challenges/KPRCA_00003/lib/printf.c
@@ -23,12 +23,13 @@
  *
  */
 #include <stdarg.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
 typedef int (*consumer_t)(void *arg, const char *buf, size_t n);
 
-int writeall(int fd, const void *buf, size_t n) {
+int writeall(intptr_t fd, const void *buf, size_t n) {
   const void *wptr = buf;
   const void *bufend = buf + n;
   while (wptr != bufend) {
@@ -219,7 +220,7 @@ error:
 }
 
 static int _consumer_fd(void *arg, const char *buf, size_t n) {
-  return writeall((int)arg, buf, n);
+  return writeall((intptr_t)arg, buf, n);
 }
 
 static int _consumer_string(void *arg, const char *buf, size_t n) {
@@ -246,7 +247,7 @@ static int _consumer_string_checked(void *arg, const char *buf, size_t n) {
   return (int)n;
 }
 
-int fdprintf(int fd, const char *fmt, ...) {
+int fdprintf(intptr_t fd, const char *fmt, ...) {
   va_list ap;
   va_start(ap, fmt);
   int ret = _printf(_consumer_fd, (void *)fd, fmt, ap);
