+++ challenges/KPRCA_00025/lib/stdlib.h
@@ -4,11 +4,12 @@
 #include <libcgc.h>
 #include <stdarg.h>
 #include <stddef.h>
+#include <stdint.h>
 
 #define isinf(x) __builtin_isinf(x)
 #define isnan(x) __builtin_isnan(x)
 
-extern int fdprintf(int fd, const char *fmt, ...);
+extern int fdprintf(intptr_t fd, const char *fmt, ...);
 extern int sprintf(char *s, const char *fmt, ...);
 #define printf(...) fdprintf(STDOUT, __VA_ARGS__)
 extern int vfdprintf(int fd, const char *fmt, va_list ap);
