+++ challenges/KPRCA_00026/lib/stdlib.h
@@ -4,13 +4,14 @@
 #include <libcgc.h>
 #include <stdarg.h>
 #include <stddef.h>
+#include <stdint.h>
 
 #include "wrapper.h"
 
 #define isinf(x) __builtin_isinf(x)
 #define isnan(x) __builtin_isnan(x)
 
-extern int fdprintf(int fd, const char *fmt, ...);
+extern int fdprintf(intptr_t fd, const char *fmt, ...);
 extern int sprintf(char *s, const char *fmt, ...);
 #define printf(...) fdprintf(STDOUT, __VA_ARGS__)
