+++ challenges/KPRCA_00039/lib/printf.c
@@ -23,6 +23,7 @@
  *
  */
 #include <stdarg.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
@@ -70,7 +71,7 @@ int output_number_printf(int fd, unsigned int x, int base, int min,
   return n + 1;
 }
 
-int fdprintf(int fd, const char *fmt, ...) {
+int fdprintf(intptr_t fd, const char *fmt, ...) {
   char *astring;
   char achar;
   int aint, i, n = 0, flags = 0, min = 0;
